fix(auth-aff): Improve author affiliation
Removes GSSP to simplify the data fetching part
wraps component in error boundary
Moves fetching deeper so we can allow for modification of the years/num_authors during an error response
Small improvements, clean-up
