+++ src/components/AuthorAffiliations/AuthorAffiliations.tsx
@@ -5,11 +5,7 @@ import {
   AlertTitle,
   Box,
   BoxProps,
-  Button,
   Checkbox,
-  Flex,
-  FormControl,
-  FormLabel,
   Heading,
   Select,
   Skeleton,
@@ -25,81 +21,32 @@ import {
   Tr,
   useBoolean,
   VisuallyHidden,
-  Wrap,
-  WrapItem,
 } from '@chakra-ui/react';
-import { isNil, pathOr, pluck } from 'ramda';
-import { isNotNilOrEmpty } from 'ramda-adjunct';
-import {
-  ChangeEventHandler,
-  Dispatch,
-  ReactElement,
-  Reducer,
-  useCallback,
-  useEffect,
-  useMemo,
-  useReducer,
-} from 'react';
-import { ExportModal } from './ExportModal';
-import { countOptions, NONESYMBOL } from './models';
+import { ReactElement, Reducer, useCallback, useReducer } from 'react';
+import { NONESYMBOL } from './models';
 import { AuthorAffStoreProvider, useAuthorAffStore } from './store';
-import { isIADSSearchParams } from '@/utils/common/guards';
-import { IADSApiSearchParams, IDocsEntity } from '@/api/search/types';
-import { useSearch } from '@/api/search/search';
-import { useAuthorAffiliationSearch } from '@/api/author-affiliation/author-affiliation';
+import { IADSApiSearchParams } from '@/api/search/types';
 import { IGroupedAuthorAffilationData } from './types';
 import { AuthorAffiliationsErrorMessage, errorMessages } from '@/components/AuthorAffiliations/ErrorMessage';
 import { parseAPIError } from '@/utils/common/parseAPIError';
+import { useFetchAffData } from '@/components/AuthorAffiliations/hooks/UseFetchAffData';
+import { AffiliationControls } from '@/components/AuthorAffiliations/AffiliationControls';
+import { useSubCaption } from '@/components/AuthorAffiliations/hooks/UseSubCaption';
 
 /** ---------- Types & constants ---------- */
 
-type AffTableState = {
+export type AffTableState = {
   maxAuthors: number;
   numYears: number;
 };
 
-type AffTableAction = { type: 'setMaxAuthors'; payload: number } | { type: 'setNumYears'; payload: number };
+export type AffTableAction = { type: 'setMaxAuthors'; payload: number } | { type: 'setNumYears'; payload: number };
 
 export type AuthorAffiliationsProps =
   | (BoxProps & { params: IAuthorAffiliationPayload; query?: IADSApiSearchParams; ssrError?: string })
   | { params?: IAuthorAffiliationPayload; query: IADSApiSearchParams };
 
 const DEFAULT_PARAMS = getAuthorAffiliationSearchParams();
-
-/** ---------- Hooks ---------- */
-
-const useBibcodesFromQuery = (query: IADSApiSearchParams) => {
-  return useSearch(query, {
-    enabled: isIADSSearchParams(query),
-    useErrorBoundary: true,
-    select: (data) => {
-      if (isNil(data)) {
-        return [];
-      }
-      const docs = pathOr<IDocsEntity[]>([], ['response', 'docs'], data);
-      return isNotNilOrEmpty(docs) ? pluck('bibcode', docs) : [];
-    },
-  });
-};
-
-const useSubCaption = (state: AffTableState) => {
-  return useMemo(() => {
-    if (!state) {
-      return null;
-    }
-    const { maxAuthors, numYears } = state;
-    const currentYear = new Date().getFullYear();
-    const parts: string[] = [];
-
-    if (isNotNilOrEmpty(numYears)) {
-      parts.push(numYears === 0 ? 'All years' : `From ${currentYear - numYears} to ${currentYear}`);
-    }
-    if (isNotNilOrEmpty(maxAuthors)) {
-      parts.push(`${maxAuthors === 0 ? 'All' : maxAuthors} author${maxAuthors !== 1 ? 's' : ''} from each work`);
-    }
-    return parts.join(' | ');
-  }, [state]);
-};
 
 /** ---------- Reducer ---------- */
 
@@ -134,113 +81,6 @@ const AffiliationHeader = ({ formState }: { formState: AffTableState }) => {
   );
 };
 
-/** ---------- Controls (left: actions, right: selects) ---------- */
-
-const AffiliationControls = ({
-  formState,
-  dispatch,
-}: {
-  formState: AffTableState;
-  dispatch: Dispatch<AffTableAction>;
-}) => {
-  const isDisabled = useAuthorAffStore((s) => s.isLoading || s.items.length === 0);
-  const reset = useAuthorAffStore((s) => s.reset);
-  const toggleAll = useAuthorAffStore((s) => s.toggleAll);
-
-  const handleYearChange: ChangeEventHandler<HTMLSelectElement> = (e) =>
-    dispatch({ type: 'setNumYears', payload: Number(e.currentTarget.value) });
-
-  const handleAuthorChange: ChangeEventHandler<HTMLSelectElement> = (e) =>
-    dispatch({ type: 'setMaxAuthors', payload: Number(e.currentTarget.value) });
-
-  return (
-    <section aria-labelledby="modify-form-area" id="author-affiliation-content">
-      <VisuallyHidden as="h3" id="modify-form-area">
-        Modify Form Parameters
-      </VisuallyHidden>
-
-      {/* Mobile: column; md+: row spaced */}
-      <Flex
-        mt="2"
-        mb="4"
-        direction={{ base: 'column', md: 'row' }}
-        align={{ base: 'stretch', md: 'center' }}
-        justify="space-between"
-        gap={3}
-      >
-        {/* Left: action buttons */}
-        <Stack direction="row" align="center">
-          <Button size="xs" variant="ghost" onClick={toggleAll} isDisabled={isDisabled}>
-            Toggle All
-          </Button>
-          <Button size="xs" variant="ghost" onClick={reset} isDisabled={isDisabled}>
-            Reset
-          </Button>
-          <ExportModal isDisabled={isDisabled} />
-        </Stack>
-
-        {/* Right: selects; Wrap keeps them tidy at small widths */}
-        <Wrap spacing="12px" justify={{ base: 'flex-start', md: 'flex-end' }}>
-          <WrapItem>
-            <FormControl minW="180px">
-              <FormLabel>Max Authors</FormLabel>
-              <Select onChange={handleAuthorChange} value={String(formState.maxAuthors)}>
-                {countOptions.map((count) => (
-                  <option value={count} key={count}>
-                    {count}
-                  </option>
-                ))}
-                <option value="0">All Authors</option>
-              </Select>
-            </FormControl>
-          </WrapItem>
-          <WrapItem>
-            <FormControl minW="160px">
-              <FormLabel>Years</FormLabel>
-              <Select onChange={handleYearChange} value={String(formState.numYears)}>
-                {countOptions.map((count) => (
-                  <option value={count} key={count}>
-                    {count}
-                  </option>
-                ))}
-                <option value="0">All Years</option>
-              </Select>
-            </FormControl>
-          </WrapItem>
-        </Wrap>
-      </Flex>
-    </section>
-  );
-};
-
-const useFetchAffData = (query: IADSApiSearchParams, formState: AffTableState) => {
-  const setItems = useAuthorAffStore((s) => s.setItems);
-  const setIsLoading = useAuthorAffStore((s) => s.setIsLoading);
-
-  const { data: bibcode } = useBibcodesFromQuery(query);
-
-  const {
-    data: items,
-    isLoading,
-    error,
-    isError,
-  } = useAuthorAffiliationSearch(
-    getAuthorAffiliationSearchParams({ maxauthor: [formState.maxAuthors], numyears: [formState.numYears], bibcode }),
-    { enabled: isNotNilOrEmpty(bibcode) },
-  );
-
-  useEffect(() => {
-    setIsLoading(isLoading);
-  }, [isLoading, setIsLoading]);
-  useEffect(() => {
-    if (isNotNilOrEmpty(items)) {
-      setItems(items);
-    }
-  }, [items, setItems]);
-
-  return { isLoading, error, isError };
-};
-
 /** ---------- Table ---------- */
 
 const AffiliationTable = (props: { query: IADSApiSearchParams; formState: AffTableState }) => {
@@ -248,7 +88,11 @@ const AffiliationTable = (props: { query: IADSApiSearchParams; formState: AffTab
   const { isLoading, error, isError } = useFetchAffData(query, formState);
   const items = useAuthorAffStore((s) => s.items);
 
-  if (isError && parseAPIError(error) !== errorMessages.noResults) {
+  if (
+    isError &&
+    // TODO: (refactor) this will be unnecessary when the service returns empty array instead of error
+    parseAPIError(error) !== errorMessages.noResults
+  ) {
     return <AuthorAffiliationsErrorMessage error={error} />;
   }
 
@@ -313,16 +157,16 @@ const SkeletonTableRows = () => {
       {[0, 1, 2, 3].map((i) => (
         <Tr key={i}>
           <Td>{i + 1}</Td>
-          <Td>
+          <Td data-testid="skeleton-row">
             <Skeleton h="3" w="20" />
           </Td>
-          <Td>
+          <Td data-testid="skeleton-row">
             <Skeleton h="3" w="20" />
           </Td>
-          <Td>
+          <Td data-testid="skeleton-row">
             <Skeleton h="3" w="40" />
           </Td>
-          <Td>
+          <Td data-testid="skeleton-row">
             <Skeleton h="3" w="20" />
           </Td>
         </Tr>
@@ -358,6 +202,7 @@ const Row = (props: { context: IGroupedAuthorAffilationData; idx: number }) => {
           onChange={() => toggle(ctx.authorName)}
           onFocus={setIsFocused.on}
           onBlur={setIsFocused.off}
+          aria-label={`select author ${ctx.authorName}`}
         >
           <Tooltip label={ctx.authorName} aria-label={ctx.authorName}>
             <Text isTruncated w="3xs">
@@ -376,6 +221,7 @@ const Row = (props: { context: IGroupedAuthorAffilationData; idx: number }) => {
               isDisabled={!state.selected}
               onFocus={setIsFocused.on}
               onBlur={setIsFocused.off}
+              aria-label={`select affiliation ${aff} for author ${ctx.authorName}`}
             >
               <Tooltip label={aff} aria-label={aff}>
                 <Text isTruncated w={['xs', 'sm']}>
@@ -403,7 +249,7 @@ const Row = (props: { context: IGroupedAuthorAffilationData; idx: number }) => {
         <VisuallyHidden as="label" htmlFor={`${ctx.authorName}-last_active_date_select`}>
           Select last active date
         </VisuallyHidden>
-        {ctx.lastActiveDate.length === 1 ? (
+        {Array.isArray(ctx.lastActiveDate) && ctx.lastActiveDate.length === 1 ? (
           <Text>{ctx.lastActiveDate[0]}</Text>
         ) : (
           <Select
@@ -414,6 +260,7 @@ const Row = (props: { context: IGroupedAuthorAffilationData; idx: number }) => {
             size="xs"
             onFocus={setIsFocused.on}
             onBlur={setIsFocused.off}
+            data-testid="last-active-date-select"
           >
             {ctx.lastActiveDate.map((date, i) => (
               <option key={`lastactivedate_${date}_${i}`} value={date}>
