+++ src/components/AuthorAffiliations/ExportModal.tsx
@@ -19,7 +19,7 @@ import {
   useDisclosure,
 } from '@chakra-ui/react';
 import { exportTypes } from './models';
-import { useExportModal } from './useExportModal';
+import { useExportModal } from './hooks/useExportModal';
 import { parseAPIError } from '@/utils/common/parseAPIError';
 import { SimpleLink } from '@/components/SimpleLink';
 
@@ -52,7 +52,7 @@ export const ExportModal = (props: ButtonProps) => {
         <ModalContent>
           <ModalCloseButton />
           <ModalHeader>
-            Exporting {numSelected} entrie{numSelected > 1 ? 's' : ''}
+            Exporting {numSelected.toLocaleString()} entr{numSelected > 1 ? 'ies' : 'y'}
           </ModalHeader>
           <ModalBody>
             <FormLabel htmlFor="export-selection">Select export format</FormLabel>
@@ -66,7 +66,7 @@ export const ExportModal = (props: ButtonProps) => {
               </Stack>
             </RadioGroup>
             <Button onClick={onFetch} mt="2" width="full" isLoading={isLoading} isDisabled={isError || noData}>
-              Export to file
+              Export
             </Button>
             {noData ? (
               <Alert status="warning" mt="2">
