+++ src/components/AuthorAffiliations/index.ts
@@ -1,2 +1,4 @@
 export * from './AuthorAffiliations';
 export * from './ErrorMessage';
+export { AffiliationControls } from '@/components/AuthorAffiliations/AffiliationControls';
+export { useSubCaption } from '@/components/AuthorAffiliations/hooks/UseSubCaption';
